telegram-mtproto-voice/requirements.txt
telethon>=1.36.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0
httpx>=0.27.0
